Remove duplication from PAPI instrumentation and CPU codegen
